FI-765 feat: add environment variable `E2ED_TERMINATION_SIGNAL`
feat: add environment variable `E2ED_TIMEOUT_FOR_GRACEFUL_SHUTDOWN_IN_SECONDS`
refactor: move `onRequest` and `onConfigureResponse` methods to thier own modules
