+++ src/utils/end/setProcessEndHandlers.ts
@@ -1,4 +1,4 @@
-import {EndE2edReason} from '../../constants/internal';
+import {e2edEnvironment, EndE2edReason} from '../../constants/internal';
 
 import {generalLog} from '../generalLog';
 
@@ -9,7 +9,13 @@ import {endE2ed} from './endE2ed';
  * @internal
  */
 const endHandler = (signal: NodeJS.Signals): void => {
-  generalLog(`Receive nodejs e2ed process end signal ${signal}`);
+  const message = `Receive nodejs e2ed process end signal ${signal}`;
+
+  e2edEnvironment.E2ED_TERMINATION_SIGNAL = signal;
+
+  // eslint-disable-next-line no-console
+  console.log(message);
+  generalLog(message);
 
   endE2ed(EndE2edReason.ProcessEndSignal);
 };
